+++ src/Visitor/Php/Symfony/FlashMessage.php
@@ -13,7 +13,6 @@ namespace Translation\Extractor\Visitor\Php\Symfony;
 
 use PhpParser\Node;
 use PhpParser\NodeVisitor;
-use Translation\Extractor\Model\SourceLocation;
 use Translation\Extractor\Visitor\Php\BasePHPVisitor;
 
 /**
@@ -48,7 +47,7 @@ final class FlashMessage extends BasePHPVisitor implements NodeVisitor
             ('add' === $name && 'getFlashBag' === $callerName && $caller instanceof Node\Expr\MethodCall)
         ) {
             if (null !== $label = $this->getStringArgument($node, 1)) {
-                $this->collection->addLocation(new SourceLocation($label, $this->getAbsoluteFilePath(), $node->getAttribute('startLine')));
+                $this->addLocation($label, $node->getAttribute('startLine'), $node);
             }
         }
     }