+++ src/Visitor/Php/Symfony/FormTypeEmptyValue.php
@@ -14,7 +14,6 @@ namespace Translation\Extractor\Visitor\Php\Symfony;
 use PhpParser\Node;
 use PhpParser\Node\Stmt;
 use PhpParser\NodeVisitor;
-use Translation\Extractor\Model\SourceLocation;
 use Translation\Extractor\Visitor\Php\BasePHPVisitor;
 
 /**
@@ -71,8 +70,7 @@ final class FormTypeEmptyValue extends BasePHPVisitor implements NodeVisitor
             return;
         }
 
-        $sl = new SourceLocation($label, $this->getAbsoluteFilePath(), $node->getAttribute('startLine'));
-        $this->collection->addLocation($sl);
+        $this->addLocation($label, $node->getAttribute('startLine'), $node);
     }
 
     public function leaveNode(Node $node)