+++ src/Visitor/Php/Symfony/FormTypeInvalidMessage.php
@@ -14,7 +14,6 @@ namespace Translation\Extractor\Visitor\Php\Symfony;
 use PhpParser\Node;
 use PhpParser\Node\Stmt;
 use PhpParser\NodeVisitor;
-use Translation\Extractor\Model\SourceLocation;
 use Translation\Extractor\Visitor\Php\BasePHPVisitor;
 
 /**
@@ -55,8 +54,7 @@ final class FormTypeInvalidMessage extends BasePHPVisitor implements NodeVisitor
                 continue;
             }
 
-            $sl = new SourceLocation($label, $this->getAbsoluteFilePath(), $node->getAttribute('startLine'), ['domain' => 'validators']);
-            $this->collection->addLocation($sl);
+            $this->addLocation($label, $node->getAttribute('startLine'), $node, ['domain' => 'validators']);
         }
     }