+++ src/Visitor/Php/Symfony/FormTypeLabelExplicit.php
@@ -14,7 +14,6 @@ namespace Translation\Extractor\Visitor\Php\Symfony;
 use PhpParser\Node;
 use PhpParser\Node\Stmt;
 use PhpParser\NodeVisitor;
-use Translation\Extractor\Model\SourceLocation;
 use Translation\Extractor\Visitor\Php\BasePHPVisitor;
 
 /**
@@ -60,8 +59,7 @@ final class FormTypeLabelExplicit extends BasePHPVisitor implements NodeVisitor
                 continue;
             }
 
-            $sl = new SourceLocation($label, $this->getAbsoluteFilePath(), $node->getAttribute('startLine'));
-            $this->collection->addLocation($sl);
+            $this->addLocation($label, $node->getAttribute('startLine'), $item);
         }
     }