+++ src/Visitor/Php/Symfony/FormTypeLabelImplicit.php
@@ -14,7 +14,6 @@ namespace Translation\Extractor\Visitor\Php\Symfony;
 use PhpParser\Node;
 use PhpParser\Node\Stmt;
 use PhpParser\NodeVisitor;
-use Translation\Extractor\Model\SourceLocation;
 use Translation\Extractor\Visitor\Php\BasePHPVisitor;
 
 /**
@@ -54,8 +53,7 @@ final class FormTypeLabelImplicit extends BasePHPVisitor implements NodeVisitor
             if (false === $customLabel) {
                 $label = $node->args[0]->value->value;
                 if (!empty($label)) {
-                    $sl = new SourceLocation($label, $this->getAbsoluteFilePath(), $node->getAttribute('startLine'));
-                    $this->collection->addLocation($sl);
+                    $this->addLocation($label, $node->getAttribute('startLine'), $node);
                 }
             }
         }